+++ m1_csrc/m1_http_client.c
@@ -18,6 +18,7 @@
 #include "main.h"
 #include "m1_http_client.h"
 #include "m1_wifi.h"
+#include "m1_esp32_hal.h"
 #include "m1_compile_cfg.h"
 #include "m1_file_browser.h"
 #include "m1_watchdog.h"
@@ -71,7 +72,7 @@ static int parse_http_headers(const char *data, int data_len,
 bool http_is_ready(void)
 {
 #ifdef M1_APP_WIFI_CONNECT_ENABLE
-	return wifi_is_connected() && get_esp32_main_init_status();
+	return wifi_is_connected() && m1_esp32_get_init_status() && get_esp32_main_init_status();
 #else
 	return false;
 #endif
@@ -209,8 +210,10 @@ http_status_t http_get(const char *url, char *response_buf, uint16_t buf_size, u
 
 	response_buf[0] = '\0';
 
-	if (!http_is_ready())
+	if (!wifi_is_connected())
 		return HTTP_ERR_NO_WIFI;
+	if (!m1_esp32_get_init_status() || !get_esp32_main_init_status())
+		return HTTP_ERR_ESP_NOT_READY;
 
 	if (!timeout_sec)
 		timeout_sec = HTTP_DEFAULT_TIMEOUT;
@@ -977,8 +980,10 @@ http_status_t http_download_to_file(const char *url, const char *sd_path,
 	if (!url || !sd_path)
 		return HTTP_ERR_INVALID_ARG;
 
-	if (!http_is_ready())
+	if (!wifi_is_connected())
 		return HTTP_ERR_NO_WIFI;
+	if (!m1_esp32_get_init_status() || !get_esp32_main_init_status())
+		return HTTP_ERR_ESP_NOT_READY;
 
 	if (!timeout_sec)
 		timeout_sec = HTTP_DEFAULT_TIMEOUT;